worked on complaint asignation
added modals to the complaints screen, modified database structure to make it possible for the admin to assign complaints to an employee, in consequence joins were implemented to get the user role in an easier way this will let the program iterate and filter over the elements. Lastly to make it more easy for the developer to see DB data, the contexts screen was heavily updated, added filters and separated each list into a separate component.

Co-Authored-By: Juan-XC7 <113059155+Juan-XC7@users.noreply.github.com>